change positioning strategy to leave gaps
This is messy and I don't intend to upstream right now: it's a
significant enough change it probably wouldn't be desirable upstream as
an optional config but I'm also not sure there's enough interest to
maintain as its own gem.

The motivation here is that when you default to consecutive positions
(1,2,3) re-ordering 1 item in large lists can lead to a lot of db
queries (if I have a 100 item list and move the last item to position
20, every record above 20 needs to get updated), and indirectly that can
lead to complications on the front-end with needing to figure out where
to order things (do you feed updates from all the records back from the
client, or track some additional state locally since you might end up
with duplicate positions?)

My solution is leaving gaps. Basically, I don't care if a list's
positions are `[1,2,3]`, `[10,20,30]`, or `[1,97,756]`. I'm not showing
the actual positions to users anywhere (and if I was it'd be easy to
generate consecutive numbers for display rather than show the raw db
value). When appending an item to a list, it gets the previous last
item's postion + a gap (configurable, 10_000 by default). When inserting
between two items, it gets the midpoint between those two items.
Prepending subtracts the gap from the first item (since the default
recommended schema here uses signed ints). Healing moves as few records
as possible.

Some also-messy notes here:

- I set up some dockerizing locally to keep the env/deps isolated
- I hardcoded credentials for the docker setup in database.yml
- I care more about postgres and issues with the mysql connection I
  changed that default.
- I increased the pool sizes in database.yml. I saw
  `test_no_duplicate_row_values` fail on timeouts, it looks like that
  test will open up to 20 simultaneous connections, so a larger pool
  size is necessary. Not sure why the checked-in config doesn't reflect
  this, maybe the usual expectation is each thread completes fast enough
  that they don't exhaust the pool and things are a bit too slow in my
  env?